requirements.txt
aniso8601==7.0.0
asgiref==3.2.3
Django==3.0.2
django-cors-headers==3.2.1
django-graphql-jwt==0.3.0
graphene==2.1.8
graphene-django==2.8.0
graphql-core==2.3.1
graphql-relay==2.0.1
promise==2.3
PyJWT==1.7.1
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.0
